Move LoginSessionView from user_api to user_authn
