+++ openedx/core/djangoapps/user_api/legacy_urls.py
@@ -38,8 +38,6 @@ urlpatterns = [
 ]
 
 urlpatterns += [
-    url(r'^v1/account/login_session/$', user_api_views.LoginSessionView.as_view(),
-        name="user_api_login_session"),
     url(r'^v1/account/password_reset/$', user_api_views.PasswordResetView.as_view(),
         name="user_api_password_reset"),
 ]
