+++ openedx/core/djangoapps/user_api/views.py
@@ -8,8 +8,7 @@ from django.db import transaction
 from django.http import HttpResponse, HttpResponseForbidden
 from django.utils.decorators import method_decorator
 from django.utils.translation import ugettext as _
-from django.views.decorators.csrf import csrf_exempt, csrf_protect, ensure_csrf_cookie
-from django.views.decorators.debug import sensitive_post_parameters
+from django.views.decorators.csrf import ensure_csrf_cookie
 from django_filters.rest_framework import DjangoFilterBackend
 from edx_rest_framework_extensions.auth.session.authentication import SessionAuthenticationAllowInactiveUser
 from opaque_keys import InvalidKeyError
@@ -25,11 +24,10 @@ from openedx.core.djangoapps.django_comment_common.models import Role
 from openedx.core.djangoapps.user_api import accounts
 from openedx.core.djangoapps.user_api.accounts.api import check_account_exists
 from openedx.core.djangoapps.user_api.api import (
-    RegistrationFormFactory,
     get_login_session_form,
     get_password_reset_form
 )
-from openedx.core.djangoapps.user_api.helpers import require_post_params, shim_student_view
+from openedx.core.lib.api.view_utils import require_post_params
 from openedx.core.djangoapps.user_api.models import UserPreference
 from openedx.core.djangoapps.user_api.preferences.api import get_country_time_zones, update_email_opt_in
 from openedx.core.djangoapps.user_api.serializers import (
@@ -40,59 +38,6 @@ from openedx.core.djangoapps.user_api.serializers import (
 from openedx.core.lib.api.permissions import ApiKeyHeaderPermission
 from student.helpers import AccountValidationError
 from util.json_request import JsonResponse
-
-
-class LoginSessionView(APIView):
-    """HTTP end-points for logging in users. """
-
-    # This end-point is available to anonymous users,
-    # so do not require authentication.
-    authentication_classes = []
-
-    @method_decorator(ensure_csrf_cookie)
-    def get(self, request):
-        return HttpResponse(get_login_session_form(request).to_json(), content_type="application/json")
-
-    @method_decorator(require_post_params(["email", "password"]))
-    @method_decorator(csrf_protect)
-    def post(self, request):
-        """Log in a user.
-
-        You must send all required form fields with the request.
-
-        You can optionally send an `analytics` param with a JSON-encoded
-        object with additional info to include in the login analytics event.
-        Currently, the only supported field is "enroll_course_id" to indicate
-        that the user logged in while enrolling in a particular course.
-
-        Arguments:
-            request (HttpRequest)
-
-        Returns:
-            HttpResponse: 200 on success
-            HttpResponse: 400 if the request is not valid.
-            HttpResponse: 403 if authentication failed.
-                403 with content "third-party-auth" if the user
-                has successfully authenticated with a third party provider
-                but does not have a linked account.
-            HttpResponse: 302 if redirecting to another page.
-
-        Example Usage:
-
-            POST /user_api/v1/login_session
-            with POST params `email`, `password`, and `remember`.
-
-            200 OK
-
-        """
-        # For the initial implementation, shim the existing login view
-        # from the student Django app.
-        from openedx.core.djangoapps.user_authn.views.login import login_user
-        return shim_student_view(login_user, check_logged_in=True)(request)
-
-    @method_decorator(sensitive_post_parameters("password"))
-    def dispatch(self, request, *args, **kwargs):
-        return super(LoginSessionView, self).dispatch(request, *args, **kwargs)
 
 
 class PasswordResetView(APIView):
